+++ Model/ContaUsuario.java
@@ -1,0 +1,8 @@
+package Model;
+
+public interface ContaUsuario {
+    Pessoa login(String usuario, String senha);
+
+    void alterarSenha(String novaSenha);
+}
+
