+++ Utils/Cores.java
@@ -1,0 +1,7 @@
+package Utils;
+
+public class Cores {
+    public static final String LAVENDER = "\u001B[38;5;183m";
+    public static final String CREME = "\u001B[38;5;229m";
+    public static final String RESET = "\u001B[0m";
+}
