--- src/Model/ContaUsuario.java
@@ -1,8 +1,0 @@
-package Model;
-
-public interface ContaUsuario {
-    Pessoa login(String usuario, String senha);
-
-    void alterarSenha(String novaSenha);
-}
-
