--- src/Model/Endereco.java
@@ -1,45 +1,0 @@
-package Model;
-
-public class Endereco {
-
-    String rua;
-    String numero;
-    String bairro;
-    String cidade;
-
-
-
-    //sets
-    public void setRua(String rua) {
-        this.rua = rua;
-    }
-
-    public void setNumero(String numero) {
-        this.numero = numero;
-    }
-
-    public void setBairro(String bairro) {
-        this.bairro = bairro;
-    }
-
-    public void setCidade(String cidade) {
-        this.cidade = cidade;
-    }
-
-    //gets
-    public String getRua() {
-        return rua;
-    }
-
-    public String getBairro() {
-        return bairro;
-    }
-
-    public String getCidade() {
-        return cidade;
-    }
-
-    public String getNumero() {
-        return numero;
-    }
-}
